Install fixes, started to rewrite translate.py webnotes/wnframework#351

+++ webnotes/core/doctype/notification_count/notification_count.py
@@ -50,6 +50,7 @@ def get_notifications():
 	}
 
 def delete_notification_count_for(doctype):
+	if webnotes.flags.in_import: return
 	webnotes.conn.sql("""delete from `tabNotification Count` where for_doctype = %s""", doctype)
 
 def delete_event_notification_count():
